Добавлена работа с АЦП STM32. Не отлажено!!!

+++ 01_f103c8t6_I2C_OLED_THERMOMETR_DS18B20_2Ch/My/Drivers/adc_ST.h
@@ -21,7 +21,7 @@
 //-----------------------------------------------------------------------------
 //Аппатартный уровень.
 void     Adc_Init(void);
-uint16_t Adc_GetMeas(uint8_t channel);
+uint32_t Adc_GetMeas(uint8_t channel);
 uint16_t Adc_GetRegDR(ADC_TypeDef *adc);
 
 uint16_t Adc_GetRegJDRx(ADC_TypeDef *adc, uint8_t ch);
